Handle metadata_hash and lease equally
This doesn't change functionality, but makes it more clear that lease
and metadata_hash are handled the same way.  Must be 32 bytes, all
zero bytes is equivalent to not being present.

We should think about how to handle the Zero address, when specified
for rekey or close_to.  I believe it should be handled similarly.